Add crypto seed API
