+++ src/integration.test.ts
@@ -1,20 +1,10 @@
-import { keyPairFromSeed, SkynetClient } from "./index";
+import { generateKeyPairAndSeed, SkynetClient } from "./index";
 
-const { publicKey, privateKey } = keyPairFromSeed(makeSeed(64));
+const { publicKey, privateKey, seed } = generateKeyPairAndSeed();
 const client = new SkynetClient("https://siasky.dev");
 
 const datakey = "HelloWorld";
 const json = { data: "thisistext" };
-
-function makeSeed(length: number) {
-  let result = "";
-  const characters = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-  const charactersLength = characters.length;
-  for (let i = 0; i < length; i++) {
-    result += characters.charAt(Math.floor(Math.random() * charactersLength));
-  }
-  return result;
-}
 
 // skip - used to verify end-to-end flow
 describe.skip("siasky.dev end to end integration test", () => {
