+++ src/utils.test.ts
@@ -7,6 +7,7 @@ import {
   uriHandshakePrefix,
   uriHandshakeResolverPrefix,
   getFileMimeType,
+  validateHnsDomain,
 } from "./utils";
 
 const portalUrl = defaultSkynetPortalUrl;
@@ -105,5 +106,12 @@ describe("getFileMimeType", () => {
 
       expect(getFileMimeType(file)).toEqual("");
     });
+  });
+});
+
+describe("validateHnsDomain", () => {
+  it("should only allow lowercase ascii character", () => {
+    expect(validateHnsDomain("domain"));
+    expect(!validateHnsDomain("domain2"));
   });
 });
